+++ App/Home/Components/Keyboard/styles.ts
@@ -1,0 +1,17 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  KeyboardContainer: {
+    display: 'flex',
+    flex: 1,
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: '#434748',
+    marginBottom: 20,
+    width: '100%',
+  },
+});
+
+export default styles;
