+++ App/Home/index.tsx
@@ -1,0 +1,44 @@
+import React, {useState} from 'react';
+import {SafeAreaView, ScrollView, Text, View} from 'react-native';
+import styles from './styles';
+import Keyboard from './Components/Keyboard';
+import {handleOnPress} from './utils';
+
+function Home() {
+  const [firstNumber, setFirstNumber] = useState<string>('');
+  const [secondNumber, setSecondNumber] = useState<string>('');
+  const [operation, setOperation] = useState<string>('');
+  const [result, setResult] = useState<number>(0);
+
+  const onPress = (toShow: string, isOperation: boolean) =>
+    handleOnPress(
+      toShow,
+      isOperation,
+      firstNumber,
+      setFirstNumber,
+      secondNumber,
+      setSecondNumber,
+      operation,
+      setOperation,
+      setResult,
+    );
+
+  return (
+    <SafeAreaView style={styles.container}>
+      <View style={styles.header}>
+        <Text
+          style={
+            styles.text
+          }>{`${firstNumber} ${operation} ${secondNumber}`}</Text>
+        <ScrollView style={styles.resultContainer} horizontal={true}>
+          <Text numberOfLines={1} style={[styles.text, styles.result]}>
+            {result}
+          </Text>
+        </ScrollView>
+      </View>
+      <Keyboard handleOnPress={onPress} />
+    </SafeAreaView>
+  );
+}
+
+export default Home;
